v 2.2

+++ src/modules/field/field.service.ts
@@ -26,9 +26,11 @@ export class Field_Service {
     private readonly fieldRepository: Field_Repository,
     private sportCenterService: SportCenterService,
     private sportCategoryService: Sport_Category_Service,
-    @Inject(forwardRef(() => Reservation_Service)) private reservationService: Reservation_Service,
-    @Inject(forwardRef(() => Field_Block_Service)) private fieldblockService: Field_Block_Service
-  ) { }
+    @Inject(forwardRef(() => Reservation_Service))
+    private reservationService: Reservation_Service,
+    @Inject(forwardRef(() => Field_Block_Service))
+    private fieldblockService: Field_Block_Service,
+  ) {}
 
   async updateField(id: string, data: UpdateFieldDto): Promise<Field> {
     const field = await this.fieldRepository.findById(id);
@@ -39,18 +41,27 @@ export class Field_Service {
   async createField(fieldData: FieldDto): Promise<Field> {
     try {
       const sportCenter: SportCenter = await this.sportCenterService.getById(
-        fieldData.sportCenterId, true
+        fieldData.sportCenterId,
+        true,
       );
 
       const sportCategory: Sport_Category | null = fieldData.sportCategoryId
         ? await this.sportCategoryService.findById(fieldData.sportCategoryId)
         : null;
 
-      const created_field: Field | undefined = await this.fieldRepository.createField(
-        sportCenter,
-        sportCategory,
-        fieldData,
-      );
+      if (!sportCenter.schedules.length) {
+        throw new ApiError(
+          ApiStatusEnum.SPORTCENTER_NEEDS_SCHEDULES_BEFORE,
+          InternalServerErrorException,
+        );
+      }
+
+      const created_field: Field | undefined =
+        await this.fieldRepository.createField(
+          sportCenter,
+          sportCategory,
+          fieldData,
+        );
 
       if (created_field === undefined) {
         throw new ApiError(
@@ -64,6 +75,7 @@ export class Field_Service {
           created_field,
           schedule,
         );
+
         if (!blocks || blocks.length === 0) {
           throw new ApiError(
             ApiStatusEnum.FIELD_BLOCK_CREATION_FAILED,
@@ -79,7 +91,8 @@ export class Field_Service {
   }
 
   async findById(id: string): Promise<Field> {
-    const found_field: Field | undefined = await this.fieldRepository.findById(id);
+    const found_field: Field | undefined =
+      await this.fieldRepository.findById(id);
 
     if (found_field === undefined) {
       throw new ApiError(ApiStatusEnum.FIELD_NOT_FOUND, NotFoundException);
@@ -111,12 +124,16 @@ export class Field_Service {
     }
   }
 
-  async deleteField(id: string): Promise<ApiResponse> {
+  async deleteField(id: string): Promise<Field> {
     try {
       const field: Field = await this.findById(id);
 
+      console.log('holaaaaaaa', field);
+
       // Actualizar el estado de todas las reservas asociadas a 'CANCELED'
       if (field.reservation && field.reservation.length > 0) {
+        console.log('holaa');
+
         await Promise.all(
           field.reservation.map(async (reservation) => {
             return await this.reservationService.cancelReservation(
@@ -124,23 +141,32 @@ export class Field_Service {
             );
           }),
         );
-
-        //TODO Se tiene que enviar un mail aca avisando a los usuarios de la cancelacion
-        const deletion_result: Field | undefined =
-          await this.fieldRepository.deleteField(field);
-
-        if (!deletion_result) {
-          throw new ApiError(
-            ApiStatusEnum.FIELD_DELETION_FAILED,
-            InternalServerErrorException,
-          );
-        }
-
-        return {
-          message: ApiStatusEnum.FIELD_DELETED_SUCCESSFULLY,
-        };
       }
+
+      if (field.reservation.length > 0) {
+        throw new ApiError(
+          ApiStatusEnum.CANT_CANCEL_RESERVATIONS,
+          InternalServerErrorException,
+        );
+      }
+
+      console.log('se va a elimianr ');
+
+      //TODO Se tiene que enviar un mail aca avisando a los usuarios de la cancelacion
+      const deletion_result: Field | undefined =
+        await this.fieldRepository.deleteField(field);
+
+      if (!deletion_result) {
+        throw new ApiError(
+          ApiStatusEnum.FIELD_DELETION_FAILED,
+          InternalServerErrorException,
+        );
+      }
+
+      return deletion_result;
     } catch (error) {
+      console.log('erro', error);
+
       throw new ApiError(error?.message, BadRequestException, error);
     }
   }
